+++ lib/widgets/theme_switch_button.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import '../features/theme/theme_cubit.dart';
+
+import '../core/shared/cubits/theme_cubit/theme_cubit.dart';
 
 class SwitchThemeButton extends StatefulWidget {
   const SwitchThemeButton({super.key});
